Reorganize runSimulations
