+++ src/utils/simulation-engine/utils/get-start-years.ts
@@ -10,7 +10,7 @@ interface Options {
 // For now, it returns every year within `market-data.json`, which represents
 // a calculation that takes into account all of history. In the future,
 // this could be more customizable based on user input.
-export default function getStartYears(duration: number, options?: Options): Array<number> {
+export default function getStartYears(duration: number, options?: Options): number[] {
   const startYears = _.chain(marketData)
     .filter(data => {
       // In this calculator, we only consider the first month of each year at this time
